requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgspec==0.20.0
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
starrail-damage-cal==4.1.2
typing-extensions==4.15.0 ; python_full_version < '3.13'
yarl==1.23.0
